% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_in_tableau.R
\name{view_in_tableau}
\alias{view_in_tableau}
\title{Open a data.frame in 'Tableau'}
\usage{
view_in_tableau(df = NULL)
}
\arguments{
\item{df}{The name of a \code{data.frame} or a \code{data.frame} itself, if not provided, an addin is launched.}
}
\description{
Open a data.frame in 'Tableau'
}
\examples{
\dontrun{

# Launch addin
view_in_tableau()

# or pass the name of a data.frame
view_in_tableau("iris")

# or directly the object
view_in_tableau(iris)

}
}
